refactor: Refactor helper/html and utils/html to remove circular dependencies
