+++ src/helper/html/index.ts
@@ -1,18 +1,11 @@
-import { escapeToBuffer, stringBufferToString } from '../../utils/html'
+import { escapeToBuffer, stringBufferToString, raw } from '../../utils/html'
 import type {
   StringBuffer,
   HtmlEscaped,
   HtmlEscapedString,
-  HtmlEscapedCallback,
 } from '../../utils/html'
 
-export const raw = (value: unknown, callbacks?: HtmlEscapedCallback[]): HtmlEscapedString => {
-  const escapedString = new String(value) as HtmlEscapedString
-  escapedString.isEscaped = true
-  escapedString.callbacks = callbacks
-
-  return escapedString
-}
+export { raw }
 
 export const html = (
   strings: TemplateStringsArray,
